requirements.txt
pyyaml>=6.0
requests>=2.32.0
watchdog>=6.0.0
